Refactor benchmark tests to different package
